+++ duoStrength.js
@@ -147,6 +147,7 @@ function retrieveOptions()
 					"showTranslationText":						true,
 						"revealHotkey":								true,
 							"revealHotkeyCode":							"Ctrl+Alt+H",
+						"showNewWords":								true,
 					"showToggleHidingTextButton":				true,
 					"showLeagues":								true,
 					"wordsButton":								true,
@@ -3678,8 +3679,24 @@ function hideTranslationText(reveal = false, setupObserver = true)
 			const hintSentence = challengeTranslatePrompt.querySelector('[data-test="hint-sentence"]');
 			
 
-			if (hintSentence.querySelectorAll(NEW_WORD_SELECTOR).length != 0)
-				return false; // There is a new word, so we don't want to be hiding this sentence.
+			if (options.showNewWords && hintSentence.querySelectorAll(NEW_WORD_SELECTOR).length != 0)
+			{
+				// There is a new word, so we don't want to be hiding this sentence.
+				hintSentence.style.filter = "none";
+				hintSentence.title = "";
+				const enableDisableButton = questionContainer.querySelector(`.hideTextEnableDisable`);
+				if (enableDisableButton !== null)
+				{
+					// Remove the enable disable button
+					const headerContainer = questionContainer.querySelector(`.hideTextEnableDisable`).parentNode;
+					const header = headerContainer.firstChild;
+					header.removeAttribute("style");
+					headerContainer.parentNode.insertBefore(header, headerContainer); // Move the header back to where it should be;
+					headerContainer.remove();
+				}
+
+				return false;
+			}
 
 			if (options.showTranslationText == false && reveal == false)
 			{
@@ -4378,8 +4395,8 @@ async function init()
 				// there is a topBarDiv so we can continue to process the page to workout what to do
 
 				// set username via the href of a link to the profile
-				let profileTabHrefParts = document.querySelector(`[data-test="profile-tab"]`).href.split("/");
-				username = profileTabHrefParts[profileTabHrefParts.length - 1];
+				let profileLinkHrefParts = document.querySelector(`[href^="/profile/"]`).href.split("/");
+				username = profileLinkHrefParts[profileLinkHrefParts.length - 1];
 
 				// topBar Div is the direct container holding the navigation butons, has class _3F_8q
 				// old method topBarDiv = dataReactRoot.childNodes[2].childNodes[1].childNodes[2].childNodes[0];